requirements.txt
vllm
nvitop
wandb
weave
langchain
scikit-learn
datasets
ipywidgets
python_dotenv